move metadata types to submodule
